refactor: 최근 본 상품 관련 파일 패키지 분리
