+++ app/src/main/java/woowacourse/shopping/view/product/catalog/ProductCatalogEventHandler.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.view.product.catalog
 
 import woowacourse.shopping.view.product.catalog.adapter.LoadMoreViewHolder
 import woowacourse.shopping.view.product.catalog.adapter.ProductViewHolder
-import woowacourse.shopping.view.product.catalog.adapter.RecentProductViewHolder
+import woowacourse.shopping.view.product.catalog.adapter.recent.RecentProductViewHolder
 
 interface ProductCatalogEventHandler :
     RecentProductViewHolder.EventHandler,
