+++ app/src/main/java/woowacourse/shopping/view/product/catalog/adapter/recent/RecentProductViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.view.product.catalog.adapter
+package woowacourse.shopping.view.product.catalog.adapter.recent
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
