+++ app/src/main/res/layout/item_recent_product.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="handler"
-            type="woowacourse.shopping.view.product.catalog.adapter.RecentProductViewHolder.EventHandler" />
+            type="woowacourse.shopping.view.product.catalog.adapter.recent.RecentProductViewHolder.EventHandler" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
